requirements.txt
requests
beautifulsoup4
pandas
streamlit
langchain
faiss-cpu
sentence-transformers
psutil
matplotlib
